+++ php-fpm/data/site/dcnt/cn/include/CClient.class.php
@@ -29,7 +29,6 @@ class CClient {
 	private static function getRec( $id ) {
 
 		CDb::open();
-
 		$rec = array();
 
 		$rs = CTable::findByID( "dcount", "dcount_id", $id, array(
@@ -88,7 +87,7 @@ class CClient {
 			"loading_threshold"=>100,
 			"rec"=>$rec,
 		);
-
+		
 		self::printLoader($cfg);
 	}
 }