+++ ChangeLog.md
@@ -1,3 +1,9 @@
 # ChangeLog for RemoteTopology
+## 0.0.4
+### Added funtion for rename and drop users(range). Also added championship editor (beta version)
+## 0.0.3 
+### Added function for delete and add user
+## 0.0.2 
+### Added complex championship, and added 'mejvuz' as champ
 ## 0.0.1
 ### Work with championship Demo2020-Code1.1 and fixed redirect